refactor(job-model): Added abstract class for common fields between DistributionJobModel and DistributionJobRequestModel.
